+++ context/app/static/js/components/ServiceStatusTable/ServiceStatusTable.jsx
@@ -12,21 +12,7 @@ import { LightBlueLink } from 'js/shared-styles/Links';
 
 import StatusIcon from './StatusIcon';
 import { useGatewayStatus } from './hooks';
-
-function buildServiceStatus(name, response, noteFunc) {
-  const { build, version, api_auth } = response;
-  const isUp = api_auth || name === 'gateway';
-  // The gateway isn't explicit: If it's not up, you wouldn't get anything at all,
-  // (and you wouldn't be able to get to the portal in the first place.)
-  return {
-    name,
-    github: build ? `https://github.com/hubmapconsortium/${name}` : undefined,
-    build,
-    version,
-    isUp,
-    note: noteFunc(response),
-  };
-}
+import { buildServiceStatus } from './utils';
 
 function ServiceStatusTable() {
   const gatewayStatus = useGatewayStatus();
@@ -40,7 +26,7 @@ function ServiceStatusTable() {
         {
           name: 'portal-ui',
           github: 'https://github.com/hubmapconsortium/portal-ui',
-          // build: TODO: pull from VERSION
+          // build: TODO: pull from VERSION, or put in package.json.
           // version,
           isUp: true,
         },
